+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# 20.12.0 - 2019-12-17
+
+Developer facing:
+
+ - Add `VuexApiRequest` `error` to the store, https://gitlab.com/gitlab-org/gitter/webapp/merge_requests/1683
+ - Add permissions `gl-group-policy-delegate`, https://gitlab.com/gitlab-org/gitter/webapp/merge_requests/1679
+ - Avoid using `authenticate bearer` for the API, https://gitlab.com/gitlab-org/gitter/webapp/merge_requests/1652
+ - Add some debug tracing for the unread banner and scrolling to chats, https://gitlab.com/gitlab-org/gitter/webapp/merge_requests/1688
+ - Remove Vue test value, https://gitlab.com/gitlab-org/gitter/webapp/merge_requests/1685
+ - Add GitLab group admin discovery, https://gitlab.com/gitlab-org/gitter/webapp/merge_requests/1681
+ - Documentation: Miscellaneous tips and tricks for developers, https://gitlab.com/gitlab-org/gitter/webapp/merge_requests/1687
+
+
 # 20.11.0 - 2019-12-10
 
  - Restore social metadata to rooms/chats, https://gitlab.com/gitlab-org/gitter/webapp/merge_requests/1665
@@ -6,8 +19,9 @@
  - Allow editing and deleting messages in Sidecar, https://gitlab.com/gitlab-org/gitter/webapp/merge_requests/1674
  - Fix archive not showing messages in the current day, https://gitlab.com/gitlab-org/gitter/webapp/merge_requests/1675
  - Remove community creation invite step, https://gitlab.com/gitlab-org/gitter/webapp/merge_requests/1677
- - Include room permissions with the Vuex seed data (fix delete message conetxt menu), https://gitlab.com/gitlab-org/gitter/webapp/merge_requests/1682
+ - Include room permissions with the VueX seed data (fix delete message context menu), https://gitlab.com/gitlab-org/gitter/webapp/merge_requests/1682
  - Edit thread message UI, https://gitlab.com/gitlab-org/gitter/webapp/merge_requests/1676
+ - Add Quote and Report actions to thread messages, https://gitlab.com/gitlab-org/gitter/webapp/merge_requests/1686
 
 Developer facing:
 
@@ -16,6 +30,7 @@ Developer facing:
  - Add GitLab API backend for groups -> `/api/v1/user/:userId/orgs` endpoint, https://gitlab.com/gitlab-org/gitter/webapp/merge_requests/1559
  - Use `ChatItemPolicy` to handle edit state, https://gitlab.com/gitlab-org/gitter/webapp/merge_requests/1673
  - Add notes about splitting from GitHub URI reservation, https://gitlab.com/gitlab-org/gitter/webapp/merge_requests/1680
+ - Add stats for threaded conversations, https://gitlab.com/gitlab-org/gitter/webapp/merge_requests/1689
 
 
 # 20.10.0 - 2019-11-27